+++ DPReduction/Utils.cpp
@@ -9,7 +9,7 @@
  *
  */
 
-#include "Utils.h"
+#include "Utils.hpp"
 
 #include <fstream>
 #include <map>
